chore: restructure core
