+++ core/server/static/default.css
