+++ core/server/static/index.html
