+++ core/server/static/index.js
