+++ core/tracker/sfv/auth_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/williamsjokvist/cfn-tracker/core/shared"
+	"github.com/williamsjokvist/cfn-tracker/core/browser"
 )
 
 // The most crucial test, to make sure authentication is always working.
@@ -15,7 +15,7 @@ func TestSFVAuthentication(t *testing.T) {
 	assert := assert.New(t)
 
 	ctx := context.Background()
-	browser, _ := shared.NewBrowser(ctx, true)
+	browser, _ := browser.NewBrowser(ctx, true)
 	sf5Tracker := NewSFVTracker(browser)
 	err := sf5Tracker.Authenticate(ctx, os.Getenv(`STEAM_USERNAME`), os.Getenv(`STEAM_PASSWORD`), true)
